Notebooks/7-Building-AI-Assistants/requirements.txt
anthropic==0.49.0
google-genai==1.5.0
google-generativeai==0.8.4
langchain-anthropic==0.3.9
langchain-community==0.3.12
langchain-core==0.3.43
langchain-openai==0.2.12
openai==1.57.4
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.43.1
matplotlib==3.10.1